moved avatar & progress-circle
