+++ apps/flows/tsconfig.json
@@ -53,7 +53,9 @@
       "@kirbydesign/designsystem/radio": ["libs/designsystem/radio/index.ts"],
       "@kirbydesign/designsystem/item-sliding": ["libs/designsystem/item-sliding/index.ts"],
       "@kirbydesign/designsystem/list": ["libs/designsystem/list/index.ts"],
-      "@kirbydesign/designsystem/form-field": ["libs/designsystem/form-field/index.ts"]
+      "@kirbydesign/designsystem/form-field": ["libs/designsystem/form-field/index.ts"],
+      "@kirbydesign/designsystem/avatar": ["libs/designsystem/avatar/index.ts"],
+      "@kirbydesign/designsystem/progress-circle": ["libs/designsystem/progress-circle/index.ts"]
     },
     "target": "es2020"
   },
