+++ libs/designsystem/avatar/ng-package.json
@@ -1,0 +1,1 @@
+{}
