+++ libs/designsystem/progress-circle/ng-package.json
@@ -1,0 +1,1 @@
+{}
