+++ libs/designsystem/progress-circle/src/progress-circle-ring.component.scss
