+++ libs/designsystem/progress-circle/src/progress-circle-ring.component.ts
