+++ libs/designsystem/progress-circle/src/progress-circle.component.scss
