+++ libs/designsystem/progress-circle/src/public_api.ts
