+++ libs/designsystem/src/lib/components/index.ts
@@ -1,7 +1,5 @@
 export * from './app';
 
-export { AvatarComponent, AvatarSize } from './avatar/avatar.component';
-
 export { KirbyBadge as BadgeComponent } from './web-component-proxies.component';
 
 export { ButtonComponent, ButtonSize } from './button/button.component';
@@ -32,8 +30,6 @@ export { SegmentedControlComponent, SegmentedControlMode} from './segmented-cont
 export { SegmentItem } from './segmented-control/segment-item';
 
 export * from './toast';
-
-export * from './progress-circle';
 
 export * from './page-local-navigation';
