+++ libs/designsystem/src/lib/index.ts
@@ -26,6 +26,8 @@ export * from '@kirbydesign/designsystem/radio';
 export * from '@kirbydesign/designsystem/list';
 export * from '@kirbydesign/designsystem/item-sliding';
 export * from '@kirbydesign/designsystem/form-field';
+export * from '@kirbydesign/designsystem/avatar';
+export * from '@kirbydesign/designsystem/progress-circle';
 
 export * from './scss/scss-helper';
