+++ libs/designsystem/tsconfig.json
@@ -28,7 +28,9 @@
       "@kirbydesign/designsystem/accordion": ["accordion/index.ts"],
       "@kirbydesign/designsystem/radio": ["radio/index.ts"],
       "@kirbydesign/designsystem/item-sliding": ["item-sliding/index.ts"],
-      "@kirbydesign/designsystem/form-field": ["form-field/index.ts"]
+      "@kirbydesign/designsystem/form-field": ["form-field/index.ts"],
+      "@kirbydesign/designsystem/avatar": ["avatar/index.ts"],
+      "@kirbydesign/designsystem/progress-circle": ["progress-circle/index.ts"]
     },
     "esModuleInterop": true,
     "target": "es2020"
