+++ libs/designsystem/tsconfig.spec.json
@@ -32,7 +32,9 @@
       "@kirbydesign/designsystem/radio": ["radio/index.ts"],
       "@kirbydesign/designsystem/item-sliding": ["item-sliding/index.ts"],
       "@kirbydesign/designsystem/list": ["list/index.ts"],
-      "@kirbydesign/designsystem/form-field": ["form-field/index.ts"]
+      "@kirbydesign/designsystem/form-field": ["form-field/index.ts"],
+      "@kirbydesign/designsystem/avatar": ["avatar/index.ts"],
+      "@kirbydesign/designsystem/progress-circle": ["progress-circle/index.ts"]
     }
   },
   "files": ["test.ts"],
