+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "designsystem",
-  "version": "7.4.1",
+  "version": "7.4.19",
   "description": "Kirby Design Angular Components. This library provides Angular wrappers for the @kirbydesign/core package, for smoother integration into Angular projects.",
   "engines": {
     "node": ">=16.0.0",
